+++ packages/picasso/src/Select/Select.tsx
@@ -694,7 +694,7 @@ export const Select = documentable(
             <OutlinedInput
               // eslint-disable-next-line react/jsx-props-no-spreading
               {...rest}
-              ref={selectRef}
+              inputRef={selectRef}
               error={error}
               disabled={disabled}
               id={id}